fix: add SSR HMR for nested server actions (#34216)
- the more complex server actions break when updated. This PR aims to
fix a number of issues such as invalidating across platforms.

<!--
Please describe the motivation for this PR, and link to relevant GitHub
issues, forums posts, or feature requests.
-->

<!--
How did you build this feature or fix this bug and why?
-->

- Was able to repro a couple complex server action cases that didn't
work and added them to an E2E test for web.