Fix incorrect file path being passed to `evalMetroAndWrapFunctions`
The `filename` here is an encoded URI. While we can call `decodeURI`
on it to reverse this and make it a valid path, this is unsafe as it's
ultimately still a path with search parameters that's been encoded.

Instead, we should use the resolved entry file path here, as it's
guaranteed to work correctly.

Failing to pass a correct file path here fails `require-from-string`
silently, which uses this to derive a base path for
Node module resolution.