[expo-constants]: fix regneration of app config (#34228)
There are scripts in both the podspec and gradle file in expo-constants
that recreates the app.config file with values from the app.json file.
These scripts are not reliably run when the app is being rebuild on
neither Android nor iOS.

Closes #33692

This fixes this by:

iOS: Added `:always_out_of_date => "1"` to the script phase defining the
script in the podspec file

Android: Added `outputs.upToDateWhen { false }` to the gradle task that
rebuilds the config file.

Rerun reproduction in original issue and change the extra-field in the
app.json file before building. Each change should be reflected in the
native code after building and running the app.

- [X] I added a `changelog.md` entry and rebuilt the package sources
according to [this short
guide](https://github.com/expo/expo/blob/main/CONTRIBUTING.md#-before-submitting)
- [X] This diff will work correctly for `npx expo prebuild` & EAS Build
(eg: updated a module plugin).

---------

Co-authored-by: Expo Bot <34669131+expo-bot@users.noreply.github.com>