fix(cli): Fix URI path being passed to `require-from-string` for static export and API route evals (#34289)
Resolves #32843

# Why

When running an Expo project inside any path with a space (or certain
other special URI-unsafe characters), Web builds would fail with an
evaluation / module resolution error at
[`require-from-string`](https://github.com/expo/expo/blob/6b3041b34f2925528aeb73302a5511e8b204f706/packages/%40expo/cli/src/start/server/getStaticRenderFunctions.ts#L131).
The debug path is set to the `MetroBundlerDevServer` URI.

This was indicative of a bug, since a URI is not always an invalid path
to be passed to `require-from-string`. The latter uses this path to
derive a base path for Node module resolution.

The bug triggers because a path, such as `~/with spaces/project/...`
would be ultimately encoded as `/Users/user/with%20spaces/project/...`.
This is now an invalid path that cannot be resolved back to a file.

# How

While it's possible to just use `resolvedEntryFilePath` as the
`fileName` (which has been renamed in some changes) to ensure that the
path is a valid module file path, modifying `createBundleUrlOsPath` to
become `createBundleOsPath` and not call `encodeURI` is safer, since it
creates a filename (with search parameters) that ensures that unique
names for various Metro options will still be generated.

This is probably relevant to keep this functioning correctly:
https://github.com/expo/expo/blob/6b3041b34f2925528aeb73302a5511e8b204f706/packages/%40expo/cli/src/start/server/metro/MetroBundlerDevServer.ts#L593

`createBundleUrlOsPath` (now: `createBundleOsPath`) was introduced in
#33543 and isn't used anywhere else, since it's purpose-written for the
modified call.

This issue was hard to debug and track down due to two issues:
- Metro raises an exception for codeframes that looks related but isn't.
The error log here should be replaced with a codeframe replacement:
https://github.com/facebook/metro/blob/ab33059a2eac88452d37ac39d6aa5cfc0f3ef691/packages/metro/src/Server.js#L1238
- The `react` require does point to the output bundle, but it's not
raising a codeframe or other debugging tools, so only the require stack
is available:
https://github.com/expo/expo/blob/6b3041b34f2925528aeb73302a5511e8b204f706/packages/%40expo/cli/src/start/server/getStaticRenderFunctions.ts#L131

## Set of changes

- Refactor `MetroBundlerDevServer` to have consistent types and track
how `filename` is used
- Switch to `resolvedEntryFilePath` (testing)
- Rename `createBundleUrlOsPath` to `createBundleOsPath` and remove
`encodeURI` from it
- Add debug messages to `evalMetroNoHandling` to make this issue more
obvious in debug logs

# Test Plan

- Reuse reproduction from linked issue by cloning it into a folder
containing spaces (e.g. `./with spaces/sdk52-web-server-error`
- Apply PR to the reproduction's changes to the `@expo/cli` dependency

### E2E

An end-to-end test has been added that uses our Playwright + Bun setups.

# Checklist

<!--
Please check the appropriate items below if they apply to your diff.
-->

- [x] I added a `changelog.md` entry and rebuilt the package sources
according to [this short
guide](https://github.com/expo/expo/blob/main/CONTRIBUTING.md#-before-submitting)
- [ ] This diff will work correctly for `npx expo prebuild` & EAS Build
(eg: updated a module plugin).
- [ ] Conforms with the [Documentation Writing Style
Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)