[android][0.77][tests]: fix testability for ErrorRecovery class in Expo Updates
To get the `enableBridgelessArchitecture` we use a toggle (by source file) depending on the version of RN to return the correct version of the ReactNativeFeatureFlags implementation.

The 0.77 implementation uses a native library to check this value which fails to load if we are in a test context.

This commit fixes this by making the enableBridgelessArchitecture value injected into the ErrorRecovery class instead of accessed from the inside.

Tests can then provide a value for this flag when testing the class.