fix: After adding support for named views some unit tests are failing.
After #34161 a few unit tests referencing a private module API that tests View Prototypes started to fail since the name of the internal JS API was changed from `ViewPrototype` to `ViewPrototypes`.

NOTE: In the review process - please verify that the change above is an internal API that has no useage outside these tests and the NativeViewManagerAdapter that contains the abstraction using this field.