[android][0.77][ci] fix/run ErrorRecovery tests on rn 0.77 (#34363)
# Why

After investigating a failing CI test with @vonovak I made the following
PR.

We have a failing ci test in Android on RN 0.77 in the `ErrorRecovery`
class inside `expo-updates`. This error is caused by the class trying to
check whether or not we're running in bridgless mode using the
`ReactNativeFeatureFlags` class that has a
`enableBridgelessArchitecture` field.

This field is implemented differently on RN 0.77 and pre 0.77. On 0.77
it uses a shared library and tries to call into native code. This [fails
here](https://github.com/facebook/react-native/blob/44da5d2ee07d06a0e11b626e2641e255f0ffde9f/packages/react-native/ReactAndroid/src/main/java/com/facebook/react/internal/featureflags/ReactNativeFeatureFlagsCxxInterop.kt#L28)
in a test context since the native library is not available.

# How

This commit fixes this by making the `ReactNativeFeatureFlags`
implementing an interface that we can pass to the constructor of the
`ErrorRecovery` class so that we can inject this feature depending on
the running context.

# Test Plan

- [X] BareExpo runs and builds
- [X] Tests are passing (At least the `ErrorRecovery` tests!)

# Checklist

- [X] I added a `changelog.md` entry and rebuilt the package sources
according to [this short
guide](https://github.com/expo/expo/blob/main/CONTRIBUTING.md#-before-submitting)
- [X] This diff will work correctly for `npx expo prebuild` & EAS Build
(eg: updated a module plugin).

---------

Co-authored-by: Expo Bot <34669131+expo-bot@users.noreply.github.com>