[chore] Remove erroneous changelog file outside of packages (#34412)
# Why

Looks like this changelog was erroneously added to in some cases https://github.com/expo/expo/commits/main/packages/%40expo/CHANGELOG.md instead of adding to the packages' changelogs themselves.

Not going to try to backfill those changelogs with the erroneous entries from this. But we should still remove this changelog so it's not confusing.

# How

rm

# Test Plan

n/a

# Checklist

<!--
Please check the appropriate items below if they apply to your diff.
-->

- [ ] I added a `changelog.md` entry and rebuilt the package sources according to [this short guide](https://github.com/expo/expo/blob/main/CONTRIBUTING.md#-before-submitting)
- [ ] This diff will work correctly for `npx expo prebuild` & EAS Build (eg: updated a module plugin).
- [ ] Conforms with the [Documentation Writing Style Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)