[android][contacts] fixed saving contacts with photos (#34432)
On Android, when saving a contact that has a photo attached to it, the
method will fail because we're not resolving the path to the photo
correctly (not using a content resolver).

Note: This was previously solved in #32097 by
[freeboub](https://github.com/freeboub) - thanks!

- Added using contentResolver to get the correct inputstream to the
photo in the Contact class
- Passed appContext from Contacts module to be able to use the
contentResolver
- Added support for testing saving an image (ie. changing picture) on
Android in BareExpo

Closes #ENG-14832

Test in BareExpo:
- Open Contacts test screen
- Click contact image and select a new image (this will save the
contact)
- Verify that the contact was successfully saved.

- [x] I added a `changelog.md` entry and rebuilt the package sources
according to [this short
guide](https://github.com/expo/expo/blob/main/CONTRIBUTING.md#-before-submitting)
- [x] This diff will work correctly for `npx expo prebuild` & EAS Build
(eg: updated a module plugin).

---------

Co-authored-by: Expo Bot <34669131+expo-bot@users.noreply.github.com>
Co-authored-by: Łukasz Kosmaty <lukasz.kosmaty@swmansion.com>