fix: [iOS][contacts] get full url for social profiles (#31771)
# Why

<!--
Please describe the motivation for this PR, and link to relevant GitHub
issues, forums posts, or feature requests.
-->

Currently with contact created with expo-contacts, when you tap on
social profiles in contact book, you are not redirected. It's due to the
url passed to `CNSocialProfile` that contains only the path

<!--
How did you build this feature or fix this bug and why?
-->



# Test Plan

<!--
Please describe how you tested this change and how a reviewer could
reproduce your test, especially if this PR does not include automated
tests! If possible, please also provide terminal output and/or
screenshots demonstrating your test/reproduction.
-->
Create a contact with a social profile and tap on it 

![RocketSim_Recording_iPhone_16_Pro_6 3_2024-09-30_11 30
23](https://github.com/user-attachments/assets/9ae0a2e4-6e72-4548-ab69-40a6f1af25cc)


# Checklist

<!--
Please check the appropriate items below if they apply to your diff.
This is required for changes to Expo modules.
-->

- [x] Documentation is up to date to reflect these changes (eg:
https://docs.expo.dev and README.md).
- [x] Conforms with the [Documentation Writing Style
Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)
- [x] This diff will work correctly for `npx expo prebuild` & EAS Build
(eg: updated a module plugin).