Include information on device v. simulator difference.
When developing identified an ambiguous issue in UX as on device my app would scan the user's face ID 3 times during each login. I was unable to replicate the issue on the simulator and was able to conclude that on iOS simulator this library does not do biometric authentication to retrieve secrets but on iOS device it does.

The documentation should mention this. Took me some gray hairs to debug and find the cause for the UX issue.