fix(cli): Add fallback module resolver for `expo` and `expo-router` dependencies (#34286)
# Why

This adds a fallback module resolution strategy to our
`withMetroMultiPlatform` resolver chain, which is intended to provide
fallbacks when a failed resolution occurs via normal Node module
resolution.

The dependencies and optional peer dependencies in `expo` and
`expo-router` are special. Several package manager issues, quirks, or
the isolated dependency strategy can cause their dependencies to either
become inaccessible or not be hoisted. There are several cases in which
this may happen.

> [!NOTE]
> Unlike the [sticky resolution
strategy](https://github.com/expo/expo/pull/31750). This does not
prevent duplicate dependencies, since it serves as a fallback. While
sticky resolution is useful for **native module** dependencies (since
only one may be built into an app at a time), it's too aggressive for
missing module resolutions that occur due to package manager bugs,
invalid hoisting trees, and isolated dependencies for non-native
modules.

<details>
<summary><strong>Reminder Note:</strong> What are isolated
dependencies?</summary>

For a reminder re. isolated dependencies. This is an installation
strategy used by `pnpm`. In this installation structure it enforces that
hoisted dependencies (and direct dependencies) are isolated to the
actual dependencies that these modules specify.

In this strategy, modules that specify dependencies are linked to a
central `node_modules/.pnpm` folder. In this folder, each dependency is
placed in its own sub folder and its dependencies are recursively linked
to other folders of this style.

This effectively means that no dependencies that aren't specified are
accessible, avoiding "implicit transitive" dependencies, which can
otherwise lead to bugs when hoisting differs due to conflicts.

</details>

### Isolated dependencies with transformations

Output ASTs are transformed using Babel. This may leave `@babel/runtime`
helpers around. `@babel/runtime` is generally an entrypoint into several
output helpers that Babel places into output bundles.

However, due to isolated dependencies, without this fallback resolvers,
users will have to install `@babel/runtime` directly and rely on
deduplication. This doesn't apply with hoisting, so it's unexpected and
a repository can fail when the package manager is changed over to `pnpm`
without the hoisted linker.

### Unhoisted dependencies

If any dependency of `expo` or `expo-router` isn't hoisted, it becomes
inaccessible to the user. While this is expected, if this is either the
result of a transformation or the result of transitive dependency rules,
the module won't be resolved.

This occurred for `expo-asset/tools/hashAssetFiles` (which will be
removed) in non-isolated dependencies cases too due to invalid hoisting.
This is an intermittent error, where the dependency tree isn't
technically incorrect, but hoisting is skipped for a dependency.

### Isolated/unhoisted peer dependencies that are auto-installed

Due to a quirk in modern package manager behaviour. Peer Dependencies
may be auto-installed but not hoisted in some cases. This depends on the
package managers' configuration.

We have several optional peer dependencies in `expo` and `expo-router`
that may become inaccessible from the project root due to this problem,
and if they're imported in the project, rather than in their own module
code, an error will occur. This is usually user error, but also
unexpected and hard to spot.

# How

The resolver created by `createExpoFallbackResolver` is added as the
last one in our chain of custom resolvers. Importantly, it's placed
after `requestPostRewrites` (previously unnamed), the prior last
resolver, which always returns a resolution (or throws).
Hence, the fallback resolver is only used when normal Node module
resolution fails.

`createExpoFallbackResolver` accepts a list of `originModuleNames` (we
pass `'expo'` and `'expo-router'`). It then:
- resolves the root of these origin modules via a regular
`require.resolve` call (meaning, all origin modules must be accessible
via `@expo/cli` and/or the project, which is expected for `expo` and
optionally `expo-router`)
- loads these modules' `package.json` and creates a regex for modules
that are either their dependencies or optional peer dependencies

The resolver it creates then activates when we cannot resolve a module
prior and:
- checks that the `moduleName` is a module specifier
- checks the `moduleName` against the regexes created above
- if a match occurs, it resolves the `moduleName` from the origin
module's installation folder instead

## Set of changes

- Create type for `StrictResolver` and `StrictResolverFactory`
- Give names to resolvers passed to `withMetroResolvers` and add their
names to the `debug` error message
- Add `createExpoFallbackResolver`

# Test Plan

- Added resolution tests to `withMetroMultiPlatform`'s tests

**Reproduction:**
- Clone https://github.com/kitten/expo-pnpm-monorepo-reproduction
- _Note:_ The reproduction is a regular pnpm workspace with an Expo
template app added to it as a workspace package. **Note that there's no
`.npmrc` containing a hoisted dependency setting.** This triggers the
related bug due to isolated dependencies.
- Run `pnpm install`
- Go to `apps/native-app/`
- Run `pnpm start --clear` and open the app (iOS or Android)
- _Note:_ This will reproduce the related build error and will show
`Unable to resolve "@babel/runtime/helpers/interopRequireDefault"...`

**Verification:**
- You can verify that `@babel/runtime` is not a direct dependency of the
app (either through `pnpm list @babel/runtime` or by checking that
`node_modules/@babel/runtime` does not exist)
- Adding `pnpm add @babel/runtime` would resolve this, since it manually
provides a resolution (If you do this, make sure to reset this before
testing the fix)

**Testing this fix:**
- Run `git checkout pr-34286 `
- _Note:_ This branch contains this PR's changes applied as a pnpm patch
(patch-package change)
- Run `pnpm install`
- Run `pnpm start --clear`

# Checklist

<!--
Please check the appropriate items below if they apply to your diff.
-->

- [x] I added a `changelog.md` entry and rebuilt the package sources
according to [this short
guide](https://github.com/expo/expo/blob/main/CONTRIBUTING.md#-before-submitting)
- [ ] This diff will work correctly for `npx expo prebuild` & EAS Build
(eg: updated a module plugin).
- [ ] Conforms with the [Documentation Writing Style
Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)