[fix][filesystem] fix deleting nested folders on android
On Android calling `delete` on a directory with children will fail.

The reason is that we're using the File.delete() method in Kotlin which only deletes the current item - and fails if the current item is a directory with children.

This commit fixes this by checking if the path is a directory and tries to delete children before deleting itself.