[fix][filesystem] fix deleting nested folders on android (#34550)
On Android calling `delete` on a directory with children will fail.

The reason is that we're using the File.delete() method in Kotlin which
only deletes the current item - and fails if the current item is a
directory with children.

Closes #34542
Closes #ENG-14882

This commit fixes this by checking if the path is a directory and tries
to delete children before deleting itself.

- Added a fixed implementation
- Added a new unit test (that failed before this fix on Android) to test
feature (succeeds on both iOS and Android now)
- Added throwing exceptions if a `delete` operation fails as iOS does.

✅ Unit tests passing

- [X] I added a `changelog.md` entry and rebuilt the package sources
according to [this short
guide](https://github.com/expo/expo/blob/main/CONTRIBUTING.md#-before-submitting)
- [X] This diff will work correctly for `npx expo prebuild` & EAS Build
(eg: updated a module plugin).

---------

Co-authored-by: Expo Bot <34669131+expo-bot@users.noreply.github.com>