[router] rework unstable_settings.initialRouteName to unstable_settings.anchor (#28644)
# Why

During normal navigation, the target screen will the initial route of a
navigator. React Navigation allows users to disable this behaviour (via
the `initial: false` option when dispatching a navigation event). When
disabled, the navigator determines the index of the target route (e.g it
might have an `initialRouteName` which will be used as the initial
route).

I'm adding this as a feature, as it has both been requested and also is
a stumbling block for new users.

# How

* Added a `initial={Boolean}` prop to `<Link />`
* `router.{navigate,push,replace}()` now accepts a second options
argument. The only option is `{ initialScreen: Boolean }`

# Test Plan

Updated unit tests

# Checklist

<!--
Please check the appropriate items below if they apply to your diff.
This is required for changes to Expo modules.
-->

- [ ] Documentation is up to date to reflect these changes (eg:
https://docs.expo.dev and README.md).
- [ ] Conforms with the [Documentation Writing Style
Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)
- [ ] This diff will work correctly for `npx expo prebuild` & EAS Build
(eg: updated a module plugin).

---------

Co-authored-by: Expo Bot <34669131+expo-bot@users.noreply.github.com>
Co-authored-by: Aman Mittal <amandeepmittal@live.com>