[expo-router] add Sitemap to exported views (#34144)
# Why

I would like to be able to import `Sitemap` to build a custom sitemap
screen in my app (e.g. show it only in devClient mode).

Currently we can do it with `Unmatched` view but not with the `Sitemap`
view.

Is there any particular reason why we shouldn't have added it too?