[ios][imagepicker] fix concurrency issue when selecting multiple items
When selecting multiple items using the image picker the implementation processed the selected images in parallell - which doesn't work well on older devices - especially when doing "disk" IO.

It was easy to reproduce this error on a low end iOS device:

1. Run BareExpo
2. Open API/ImagePicker
3. Run `requireMediaLIbraryPermissionsAsync`
4. Under `launchImageLibraryAsync` ensure that `allowsMultipleSelection` is checked
5. Run `launchImageLibraryAsync` and select more than one image
6. Press "Add" button in the image picker

Observe: On low-end devices the function never returns and the app freezes. Inspection in XCode reveals that we have multiple threads running at once that are stuck waiting for each other.

Expected: All images are visible after selection

This commit fixes the above issue by using the `asyncMap` method instead of the `concurrentMap` that was previously used.

I also added a note about it in the comments in the `concurrentMap` function.

Tested on iPhone S which was not able to load even two images from the library at once - now it loads 8 images without any problem. The experience is also that it is fast enough.