[android] fix: runtimeVersion not visible in dev-client #34090 (#34398)
wasn't displaying correctly under the old architecture on Android - it
is fine and visible on the new architecture on both Android and iOS.

This was eventually confirmed after some back and forth I ended up with
the following reasoning:

The field `runtimeVersion` comes from the `expo-updates` package and is
read in `DevLauncherInternalModule.getUpdatesConfig()` method using the
following statement:

```kt
private fun getUpdatesConfig(): WritableMap {
  // ...
  val runtimeVersion = controller.updatesInterface?.runtimeVersion
  // ...
}
```

The problem was that the dependency chain required a specific order of
initialisation to ensure that both `controller` and
`controller.updatesInterface` was correctly set before this would work.

The `controller` field is set in the
`UpdatesController.initialiseWithoutStarting` method:

```kt
@JvmStatic fun initializeWithoutStarting(context: Context) {
  // ...
  if (BuildConfig.USE_DEV_CLIENT) {
    val devLauncherController = initializeAsDevLauncherWithoutStarting(context)
    singletonInstance = devLauncherController
    UpdatesControllerRegistry.controller = WeakReference(devLauncherController) // <----- HERE
  } else {
    singletonInstance = DisabledUpdatesController(context, null)
  }
  // ...
}
```

This code is called after we initialise the `DevLauncherController` and
initialize the instance's `updatesInterface` member:

```kt
fun createApplicationLifecycleListeners(context: Context?): List<ApplicationLifecycleListener> =
  listOf(
    object : ApplicationLifecycleListener {
      override fun onCreate(application: Application?) {
        check(application is ReactApplication)
        DevLauncherController.initialize(application)
        UpdatesControllerRegistry.controller?.get()?.let {
          DevLauncherController.instance.updatesInterface = it  // <----- HERE
          it.updatesInterfaceCallbacks = WeakReference(DevLauncherController.instance)
        }
      }
    }
  )
```

For this to work the instance of the `DevLauncherController` has to be
set - which is done as a sideeffect in
`DevLauncherController.initialize(application)`.

*The problem is that this side effect was not called on the old
architecture.*

The side effect is called from within the `ReactHostWrapper.kt` when
calling the `reactHostProvider()` function argument on initialisation.
When using the old architecture this was not done as a solution to a
problem with multiple reloads that was addressed in #32532.

This commit proposes a solution to this where the
`DevLauncherController` making sure the necessary side effects is run at
the correct time:

```kt
@JvmStatic
fun initialize(reactApplication: ReactApplication, additionalPackages: List<ReactPackage>? = null, launcherClass: Class<*>? = null) {
  // ...
  val reactHost = reactApplication.reactHost
  initialize(reactApplication as Context, ReactHostWrapper(reactApplication.reactNativeHost) { reactHost })
  // ...
}
```

Here we're calling the `reactHost` property *only once* and then pass a
new lambda to the ReactHostWrapper so it can initialize without having
to consider these side effects.

Closes #ENG-14823

- We should be able to see the `runtimeVersion` field in the dev-client
as noted in the original issue when running with the old architecture.
- We should be able to reload on the old architecture without the app
reloading multiple times (#32532)

- [X] I added a `changelog.md` entry and rebuilt the package sources
according to [this short
guide](https://github.com/expo/expo/blob/main/CONTRIBUTING.md#-before-submitting)
- [X] This diff will work correctly for `npx expo prebuild` & EAS Build
(eg: updated a module plugin).

---------

Co-authored-by: Expo Bot <34669131+expo-bot@users.noreply.github.com>