[android][imagepicker] duplicate files stored when picking images from camera
When picking an image using the `launchCameraAsync` method, inspecting the device's file explorer reveals that two files are created in the cache directory - not only the one returned to the caller. This makes it hard for the caller to control deletion of the file and might cause the cache folder to grow uncontrolled as described in #25728.

This commit fixes this by explicitly deleting the temporary file created as a target for the camera instead of just leaving it.

Closes #25728