[video][android][ios] Add support for listing available video tracks (#33458)
# Why

There were requests for the `naturalSize` prop from `expo-av`, which is
currently missing in `expo-video`. In `expo-av` it returned the
resolution being currently played back in the viewport. I have decided
that this is not enough, as each video can be a HLS source having
multiple available resolutions, frameRates, bitrates etc.

# How

- I created a `VideoTrack` record with all necessary information about a
video track. For Android it will be very useful for selecting a specific
track in the future, on iOS it won't be possible but the `bitrate`
property will be useful, as it is possible to partially control the
track being played back by capping the available resolution.
- Introduced `sourceLoad` (the naming is debatable, I'm very open to
suggestions), which contains all information that should be available
once a video track is loaded. This will also make the
`availableSubtitleTracksChange` event redundant, so we can mark is as
deprecated for SDK53 and remove it in a future SDK
- On Android it was relatively easy to create the VideoTracks
- On iOS for regular sources it's also easy, but for HLS it was a bit of
a hell to do it. For HLS we:
- We fetch the playlist file and parse it to obtain information about
all available video tracks.
- When a video track changes (AVKit sends an event but doesn't provide
any data) we can parse the last `AVPlayerItem` log entry to obtain a uri
to which it connected.
- The uri contains an id, which we previously parsed while saving the
playlist file, therefore we can associate it with a VideoTrack.

We now have SubtitleTracks and VideoTracks. In the future we will also
add AudioTracks, to finish the "trio" and allow for switching the
current audio language.

# Test Plan

Tested in BareExpo on Android and iOS on HLS and non-HLS sources


https://github.com/user-attachments/assets/6f918f27-4f65-49b0-bc71-3efcb0868c32