[ios][location] add error handler to streaming location/device methods
# Why

When using the streaming location/heading methods (`watchHeadingAsync` / `watchPositionAsync`) from JS, we're not handling errors very well.

It is possible for the location manager to stop being able to provide location while we're streaming. This can be tested by changing the `Features` -> `Location` to `None` while streaming location using f.ex. `City ride` in the simulator. This is a use case that can happen on a real device as well.

If this happens, an exception is thrown on the native side. This exception is not handled by the JS code in any way since all the location/heading info is delievered through an event emitter.

# How

This commit fixes the above by adding an optional errorHandler callback for the above methods, and by implementing a catch clause on the native side that will emit an error that will be caught in the JS code and passed back to the caller.

In addition there was an error with the streamer (location/heading) in swift causing them to not stop emitting after we close the subscription. The solution here was to do as on Android, to explicitly stop the streamers when we unsubscribe (on the native side).

I also refactored the Location methods to better handle errors, both when initially starting streaming, but also when requesting heading through `getHeadingAsync`.

closes #34714

# Testing