[video][android][ios] Add video caching (#31781)
# Why

Adds video caching for regular video files for Android and iOS. On
Android HLS videos also work.

# How

On Android used `SimpleCache` to implement the cache. On iOS used a
`AVAssetResourceLoaderDelegate` to capture the requests, fulfill them
and cache the responses.

The iOS structure
- `ResourceLoaderDelegate` - implements AVAssetResourceLoaderDelegate,
URLSessionDelegate - recieves requests from the AVPlayer and fulfills
them using an URL session. When receiving a request checks if the data
is available in the CachedResource and fully/partially fulfills it from
the cache.
- `CachedResource` - Meant to represent a whole of a cached video
(metadata + video data).
- `CachableRequest `- Holds a loadingRequest and associated variables
for each loadingRequest received from the AVPlayer. Holds the data
received from the server until the loadingRequest is fulfilled.
- `MediaInfo` - Holds information necessary to play the video from a
file, including the mimeType, headers received from the server response
and loadedDataRanges - ranges of data that are already available in the
file.

# Test Plan

Tested in BareExpo on Android and iOS