[video][ios] Fix safe area insets not updating for native controls (#32864)
# Why

Safe area insets for the native controls are not updating when view
moves from underneath the notch.
Fixes https://github.com/expo/expo/issues/30722

# How

Forces re-update of the native controls when the insets for a view
change. The solution is kinda ugly, but nothing else I've tried has
worked.

# Test Plan

Tested in BareExpo and the repro app from the issue.