[router] Fix getPathFromState generating invalid path for index routes (#34668)
# Why

When index routes are hoisted `getPathFromState` will generate an
invalid path. This can be demonstrated in this repo where the Tabs get
the incorrect path.
https://github.com/EvanBacon/expo-router-forms-components.

This issue only surfaces on components that call getPathFromState to
generate a URL. The location bar URL does not demostrate this issue, but
thats because during partial navigation a second navigation event
occurs. When you navigate to a partial state, it will load the navigator
which will "fix" the remaining state - and cause the location URL to
correct itself. Hence this issue only occurs in places where the URL is
"snapshotted", like the TabBar

The forked version of `getPathFromState` where it handles routing to
hoisted routes with incomplete pathnames (e.g routing to `/(a)/(b)` with
an expected screen of `./(a)/(b)/hoisted/route`) was not checking for a
valid index route.

# Test Plan

Added unit tests. Tested with repro above

---------

Co-authored-by: Expo Bot <34669131+expo-bot@users.noreply.github.com>