[core][ios] Re-Dispatch `View` `AsyncFunction` calls when the `View` is not registered (#35176)
# Why

Fixes https://github.com/expo/expo/issues/33582
Similar to
https://github.com/expo/expo/commit/bfe63e33cef041d1b3a1004fed0abefd54504bda

In iOS on new Architecture when a view function is called from the
initial render
```
useEffect({ viewRef.current.someAsyncFunction() },[])
```
it will be called before the view is added to the viewRegistry, which
means that the argument conversion will fail (the first argument of the
function is the view)

# How

We check for the conditions in which this situation occurs and
re-dispatch the function up to three times waiting for the viewManager
to get the information about the view. This is sub-optimal, but we don't
have access to the scheduler, which would allow us to do this properly.

# Test Plan

Tested in BareExpo on iOS