chore(docs): slightly clean up submission docs (#35190)
# Why

- I want to help people understand "going to the iOS App Store is easy
and takes ~25 hours for the first go, then ~20 minutes and a single
command for all subsequent runs". No expo user should be confused about
this process, they should have full confidence that the app they wrote
can get to TestFlight in a single command.
- This PR gets the ball rolling but it seems like we need to make some
larger changes. I'd like the submit story to be split by app store and
not try to be so cross-platform since the process is very different
across both.
-
https://exponent-internal.slack.com/archives/C1QMWQ1P0/p1740601632657859?thread_ts=1740598578.432709&cid=C1QMWQ1P0

---------

Co-authored-by: Aman Mittal <amandeepmittal@live.com>