[video][android] Use `ExoPlayer.switchTargetView` for changing the `SurfaceView` of the player (#34692)
# Why

From ExoPlayer javadoc of the `setPlayer` method:

> To transition a Player from targeting one view to another, it's
recommended to use switchTargetView(Player, PlayerView, PlayerView)
rather than this method. If you do wish to use this method directly, be
sure to attach the player to the new view before calling setPlayer(null)
to detach it from the old one. This ordering is significantly more
efficient and may allow for more seamless transitions.
> Params:
> player – The Player to use, or null to detach the current player. Only
players which are accessed on the main thread are supported (player.
getApplicationLooper() == Looper. getMainLooper()).

Also noticed that once a VideoView is destroyed with a mounted player
the `VideoManager` will keep a reference to it because of not being
removed from `videoPlayersToVideoViews`. Included the fix in this PR
since it's a one-line fix (setting ` it.videoPlayer = null` in
`OnViewDestroy`

# How

Used `switchPlayerView` for changing the player view, as well as ensured
that no references to the view are kept after it has been destroyed.

# Test Plan

Tested in BareExpo on Android